changes made on admin side

+++ app/Http/Controllers/Admin/EventsController.php
@@ -12,21 +12,90 @@ use Gate;
 use Illuminate\Http\Request;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use Symfony\Component\HttpFoundation\Response;
+use Yajra\DataTables\Facades\DataTables;
 
 class EventsController extends Controller
 {
     use MediaUploadingTrait;
 
-    public function index()
+    public function index(Request $request)
     {
+        abort_if(Gate::denies('event_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $events = Event::with(['media'])->get();
+        if ($request->ajax()) {
+            $query = Event::query()->select(sprintf('%s.*', (new Event())->table));
+            $table = Datatables::of($query);
 
-        return view('admin.events.index', compact('events'));
+            $table->addColumn('placeholder', '&nbsp;');
+            $table->addColumn('actions', '&nbsp;');
+
+            $table->editColumn('actions', function ($row) {
+                $viewGate = 'event_show';
+                $editGate = 'event_edit';
+                $deleteGate = 'event_delete';
+                $crudRoutePart = 'events';
+
+                return view('partials.datatablesActions', compact(
+                'viewGate',
+                'editGate',
+                'deleteGate',
+                'crudRoutePart',
+                'row'
+            ));
+            });
+
+            $table->editColumn('id', function ($row) {
+                return $row->id ? $row->id : '';
+            });
+            $table->editColumn('name', function ($row) {
+                return $row->name ? $row->name : '';
+            });
+            $table->editColumn('event_images', function ($row) {
+                if (!$row->event_images) {
+                    return '';
+                }
+                $links = [];
+                foreach ($row->event_images as $media) {
+                    $links[] = '<a href="' . $media->getUrl() . '" target="_blank"><img src="' . $media->getUrl('thumb') . '" width="50px" height="50px"></a>';
+                }
+
+                return implode(' ', $links);
+            });
+
+            $table->editColumn('location', function ($row) {
+                return $row->location ? $row->location : '';
+            });
+            $table->editColumn('event_type', function ($row) {
+                return $row->event_type ? Event::EVENT_TYPE_RADIO[$row->event_type] : '';
+            });
+            $table->editColumn('reporting_time', function ($row) {
+                return $row->reporting_time ? $row->reporting_time : '';
+            });
+            $table->editColumn('start_time', function ($row) {
+                return $row->start_time ? $row->start_time : '';
+            });
+            $table->editColumn('end_time', function ($row) {
+                return $row->end_time ? $row->end_time : '';
+            });
+            $table->editColumn('is_active', function ($row) {
+                return $row->is_active ? Event::IS_ACTIVE_RADIO[$row->is_active] : '';
+            });
+            $table->editColumn('is_cancelled', function ($row) {
+                return $row->is_cancelled ? Event::IS_CANCELLED_RADIO[$row->is_cancelled] : '';
+            });
+
+            $table->rawColumns(['actions', 'placeholder', 'event_images']);
+
+            return $table->make(true);
+        }
+
+        return view('admin.events.index');
     }
 
     public function create()
     {
+        abort_if(Gate::denies('event_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.events.create');
     }
 
@@ -47,6 +116,7 @@ class EventsController extends Controller
 
     public function edit(Event $event)
     {
+        abort_if(Gate::denies('event_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('admin.events.edit', compact('event'));
     }
@@ -74,12 +144,14 @@ class EventsController extends Controller
 
     public function show(Event $event)
     {
+        abort_if(Gate::denies('event_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('admin.events.show', compact('event'));
     }
 
     public function destroy(Event $event)
     {
+        abort_if(Gate::denies('event_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $event->delete();
 
@@ -95,6 +167,7 @@ class EventsController extends Controller
 
     public function storeCKEditorImages(Request $request)
     {
+        abort_if(Gate::denies('event_create') && Gate::denies('event_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $model         = new Event();
         $model->id     = $request->input('crud_id', 0);
